feat(messaging): move HogVM bytecode combination to coordinator
Move combine_filter_bytecodes from backfill workflow to filter_storage
to optimize performance by combining bytecodes once during filter storage
instead of on every execution.

Changes:
- Move combine_filter_bytecodes function to filter_storage.py
- Update store_filters() to pre-compute and store combined bytecode
- Update get_filters_and_properties() to return combined bytecode
- Simplify backfill workflow to use pre-computed bytecode
- Add backward compatibility for missing combined_bytecode
- Update tests to handle new 3-tuple return format
