+++ posthog/temporal/messaging/filter_storage.py
@@ -14,11 +14,14 @@ TTL Considerations:
 
 import json
 import hashlib
+from typing import Any
 
 import structlog
 
 from posthog.redis import get_client
 from posthog.temporal.messaging.types import PersonPropertyFilter
+
+from common.hogvm.python.operation import HOGQL_BYTECODE_IDENTIFIER, Operation
 
 KEY_PREFIX = "backfill_person_properties_filters:"
 # TTL sizing: Worst case workflow duration ~48h (3 × 12h retries + batch delays)
@@ -26,6 +29,32 @@ KEY_PREFIX = "backfill_person_properties_filters:"
 DEFAULT_TTL = 72 * 60 * 60  # 3 days
 
 logger = structlog.get_logger(__name__)
+
+
+def combine_filter_bytecodes(filters: list[PersonPropertyFilter]) -> list[Any]:
+    """Combine multiple filter bytecodes into a single bytecode returning a dict.
+
+    Strips the header from each filter's bytecode body, interleaves condition_hash
+    keys, and appends a DICT opcode to collect all results into
+    {condition_hash: bool_result, ...}.
+
+    Built once at coordinator startup and reused for every person.
+    """
+    combined: list[Any] = [HOGQL_BYTECODE_IDENTIFIER, 1]
+
+    valid_count = 0
+    for f in filters:
+        if len(f.bytecode) <= 2:
+            continue
+        combined.append(Operation.STRING)
+        combined.append(f.condition_hash)
+        combined.extend(f.bytecode[2:])
+        valid_count += 1
+
+    combined.append(Operation.DICT)
+    combined.append(valid_count)
+
+    return combined
 
 
 def store_filters(filters: list[PersonPropertyFilter], team_id: int, ttl: int = DEFAULT_TTL) -> str:
@@ -57,10 +86,14 @@ def store_filters(filters: list[PersonPropertyFilter], team_id: int, ttl: int = 
         if f.property_key:
             person_properties.add(f.property_key)
 
-    # Create storage object containing both filters and properties
+    # Combine all filter bytecodes into a single optimized bytecode
+    combined_bytecode = combine_filter_bytecodes(filters)
+
+    # Create storage object containing filters, properties, and combined bytecode
     storage_data = {
         "filters": filter_data,
         "person_properties": sorted(person_properties),  # Sort for consistent ordering
+        "combined_bytecode": combined_bytecode,
     }
 
     # Create hash of the storage data for the key
@@ -74,15 +107,15 @@ def store_filters(filters: list[PersonPropertyFilter], team_id: int, ttl: int = 
     return storage_key
 
 
-def get_filters_and_properties(storage_key: str) -> tuple[list[PersonPropertyFilter], list[str]] | None:
+def get_filters_and_properties(storage_key: str) -> tuple[list[PersonPropertyFilter], list[str], list[Any]] | None:
     """
-    Retrieve both filters and person properties using a storage key.
+    Retrieve filters, person properties, and combined bytecode using a storage key.
 
     Args:
         storage_key: Key returned by store_filters
 
     Returns:
-        Tuple of (filters, person_properties), or None if not found
+        Tuple of (filters, person_properties, combined_bytecode), or None if not found
 
     Note:
         If workflows consistently exceed 48h and TTL becomes an issue,
@@ -95,6 +128,7 @@ def get_filters_and_properties(storage_key: str) -> tuple[list[PersonPropertyFil
     storage_data = json.loads(data.decode("utf-8"))
     filter_data = storage_data["filters"]
     person_properties = storage_data["person_properties"]
+    combined_bytecode = storage_data.get("combined_bytecode")
 
     # Reconstruct PersonPropertyFilter objects
     filters = [
@@ -107,4 +141,8 @@ def get_filters_and_properties(storage_key: str) -> tuple[list[PersonPropertyFil
         for item in filter_data
     ]
 
-    return filters, person_properties
+    # If combined_bytecode is not present (backward compatibility), generate it
+    if combined_bytecode is None:
+        combined_bytecode = combine_filter_bytecodes(filters)
+
+    return filters, person_properties, combined_bytecode
